cannong: rootdir: Android.mk: Add entry for fstab on ramdisk root
Devices with Android Dynamic Partitions need a fstab on the
ramdisk root because they don't support system-as-root and
because it is not possible to specify the dynamic partitions
to mount early in the kernel device-tree.

For this reason, it is critical to have a fstab file in the
ramdisk, since it's the only place where we can read it as
per the specifications of the Android Open Source Project (AOSP):
add an entry to the makefile so that we can select this "module".

Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>